Delete account

+++ lib/modules/menu/profile_page.dart
@@ -1,5 +1,8 @@
 import 'package:cashflow/modules/auth/auth_controller.dart';
+import 'package:cashflow/shared/components/alert_dialog_info_component.dart';
+import 'package:cashflow/shared/models/action_model.dart';
 import 'package:cashflow/shared/theme/constants/app_colors.dart';
+import 'package:cashflow/shared/util/enums.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
@@ -98,7 +101,33 @@ class ProfilePage extends StatelessWidget {
                       ),
                     ),
                     onTap: () async {
-                      await _authController.deleteUser();
+                      await showDialog(
+                        context: context,
+                        builder: (BuildContext context) {
+                          return AlertDialogInfoComponent(
+                            alertDialogType: AlertDialogType.important,
+                            title: 'Excluir conta',
+                            description:
+                                'Você tem certeza que deseja excluir sua conta? Esta ação não pode ser desfeita.',
+
+                            actions: [
+                              ActionModel(
+                                title: "Excluir",
+                                onTap: () async {
+                                  await _authController.deleteUser();
+                                  Modular.to.navigate('/auth/');
+                                },
+                              ),
+                              ActionModel(
+                                title: "Cancelar",
+                                onTap: () {
+                                  Modular.to.pop();
+                                },
+                              ),
+                            ],
+                          );
+                        },
+                      );
                     },
                   ),
                 ),
